Fix a Windows burn from the last commit. bugref:9376: Complete hardware cursor implementation in VMSVGA.
svn:sync-xref-src-repo-rev: r128592

+++ src/VBox/Main/src-client/DisplayImpl.cpp
@@ -1,4 +1,4 @@
-/* $Id: DisplayImpl.cpp 77130 2019-02-01 16:55:20Z noreply@oracle.com $ */
+/* $Id: DisplayImpl.cpp 77133 2019-02-01 17:46:58Z noreply@oracle.com $ */
 /** @file
  * VirtualBox COM class implementation
  */
@@ -4453,7 +4453,7 @@ DECLCALLBACK(void) Display::i_displayVBVAInputMappingUpdate(PPDMIDISPLAYCONNECTO
     pThis->i_handleUpdateVBVAInputMapping(xOrigin, yOrigin, cx, cy);
 }
 
-DECLCALLBACK(void) Display::i_displayVBVAReportCursorPosition(PPDMIDISPLAYCONNECTOR pInterface, uint32_t fFlags, ULONG aScreenId, uint32_t x, uint32_t y)
+DECLCALLBACK(void) Display::i_displayVBVAReportCursorPosition(PPDMIDISPLAYCONNECTOR pInterface, uint32_t fFlags, uint32_t aScreenId, uint32_t x, uint32_t y)
 {
     LogFlowFunc(("\n"));
